# Implement more efficient pack and unpack uint5