Summary:
Implemented more efficient 5-bit packing and unpacking algorithm, for 8/64/128 values.
The algorithm is commented in code, but you may also refer to T204077841 for discussion.

## Before
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
benchmark_pack_uint_values<1>/128/8           17.3 ns         17.2 ns     40530134
...
benchmark_pack_uint_values<5>/128/8           36.8 ns         36.5 ns     18974458
benchmark_pack_uint_values<5>/128/64          5.47 ns         5.43 ns    128341462
benchmark_pack_uint_values<5>/128/128         2.91 ns         2.70 ns    261633340
benchmark_unpack_uint_values<5>/128/8         28.8 ns         28.6 ns     24475696
benchmark_unpack_uint_values<5>/128/64        6.14 ns         5.65 ns    124953143
benchmark_unpack_uint_values<5>/128/128       2.90 ns         2.88 ns    242818639
```

## After
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
benchmark_pack_uint_values<1>/128/8           17.9 ns         17.5 ns     40221794
...
benchmark_pack_uint_values<5>/128/8           24.9 ns         24.8 ns     28330676
benchmark_pack_uint_values<5>/128/64          2.63 ns         2.61 ns    267856460
benchmark_pack_uint_values<5>/128/128         2.04 ns         2.03 ns    344166380
benchmark_unpack_uint_values<5>/128/8         22.1 ns         22.0 ns     31850032
benchmark_unpack_uint_values<5>/128/64        2.92 ns         2.89 ns    242508228
benchmark_unpack_uint_values<5>/128/128       2.33 ns         2.25 ns    310688575
```

Reviewed By: metascroy

Differential Revision: D64703548